refactor to expose the comparison through the makefile
